Refactor entryParser tests
